+++ primefaces/src/main/java/org/primefaces/component/radarchart/RadarChartBase.java
@@ -23,43 +23,17 @@
  */
 package org.primefaces.component.radarchart;
 
-import javax.faces.component.UIComponentBase;
-import javax.faces.component.behavior.ClientBehaviorHolder;
+import org.primefaces.component.api.UIChart;
 
-import org.primefaces.component.api.PrimeClientBehaviorHolder;
-import org.primefaces.component.api.Widget;
-
-public abstract class RadarChartBase extends UIComponentBase implements Widget, ClientBehaviorHolder, PrimeClientBehaviorHolder {
-
-    public static final String COMPONENT_FAMILY = "org.primefaces.component";
+public abstract class RadarChartBase extends UIChart {
 
     public static final String DEFAULT_RENDERER = "org.primefaces.component.RadarChartRenderer";
-
-    public enum PropertyKeys {
-
-        widgetVar,
-        model,
-        style,
-        styleClass
-    }
 
     public RadarChartBase() {
         setRendererType(DEFAULT_RENDERER);
     }
 
     @Override
-    public String getFamily() {
-        return COMPONENT_FAMILY;
-    }
-
-    public String getWidgetVar() {
-        return (String) getStateHelper().eval(PropertyKeys.widgetVar, null);
-    }
-
-    public void setWidgetVar(String widgetVar) {
-        getStateHelper().put(PropertyKeys.widgetVar, widgetVar);
-    }
-
     public org.primefaces.model.charts.radar.RadarChartModel getModel() {
         return (org.primefaces.model.charts.radar.RadarChartModel) getStateHelper().eval(PropertyKeys.model, null);
     }
@@ -68,19 +42,4 @@ public abstract class RadarChartBase extends UIComponentBase implements Widget, 
         getStateHelper().put(PropertyKeys.model, model);
     }
 
-    public String getStyle() {
-        return (String) getStateHelper().eval(PropertyKeys.style, null);
-    }
-
-    public void setStyle(String style) {
-        getStateHelper().put(PropertyKeys.style, style);
-    }
-
-    public String getStyleClass() {
-        return (String) getStateHelper().eval(PropertyKeys.styleClass, null);
-    }
-
-    public void setStyleClass(String styleClass) {
-        getStateHelper().put(PropertyKeys.styleClass, styleClass);
-    }
 }