+++ primefaces/src/main/java/org/primefaces/component/scatterchart/ScatterChartRenderer.java
@@ -38,11 +38,7 @@ public class ScatterChartRenderer extends ChartRenderer {
     @Override
     public void encodeEnd(FacesContext context, UIComponent component) throws IOException {
         ScatterChart chart = (ScatterChart) component;
-        String clientId = chart.getClientId(context);
-        String style = chart.getStyle();
-        String styleClass = chart.getStyleClass();
-
-        encodeMarkup(context, clientId, style, styleClass);
+        encodeMarkup(context, chart);
         encodeScript(context, chart);
     }