+++ primefaces/src/main/java/org/primefaces/component/selectmanymenu/SelectManyMenuRenderer.java
@@ -99,7 +99,8 @@ public class SelectManyMenuRenderer extends SelectManyRenderer {
             wb.attr("filter", true)
                     .attr("filterMatchMode", menu.getFilterMatchMode(), null)
                     .nativeAttr("filterFunction", menu.getFilterFunction(), null)
-                    .attr("caseSensitive", menu.isCaseSensitive(), false);
+                    .attr("caseSensitive", menu.isCaseSensitive(), false)
+                    .attr("filterNormalize", menu.isFilterNormalize(), false);
         }
 
         wb.finish();