+++ primefaces/src/main/java/org/primefaces/renderkit/MenuItemAwareRenderer.java
@@ -41,11 +41,13 @@ import javax.faces.event.PhaseId;
 import org.primefaces.behavior.confirm.ConfirmBehavior;
 import org.primefaces.component.api.*;
 import org.primefaces.component.divider.Divider;
+import org.primefaces.component.menuitem.UIMenuItem;
 import org.primefaces.event.MenuActionEvent;
 import org.primefaces.model.menu.*;
 import org.primefaces.util.ComponentTraversalUtils;
 import org.primefaces.util.Constants;
 import org.primefaces.util.HTML;
+import org.primefaces.util.LangUtils;
 
 public class MenuItemAwareRenderer extends OutcomeTargetRenderer {
 
@@ -61,9 +63,15 @@ public class MenuItemAwareRenderer extends OutcomeTargetRenderer {
         setConfirmationScript(context, menuitem);
 
         String onclick = menuitem.getOnclick();
+        boolean isLink = LangUtils.isNotBlank(menuitem.getUrl()) || LangUtils.isNotBlank(menuitem.getOutcome());
+        boolean isCommand = LangUtils.isNotBlank(menuitem.getCommand());
+        if (!isCommand && menuitem instanceof UIMenuItem) {
+            UIMenuItem uim = (UIMenuItem) menuitem;
+            isCommand = uim.getActionExpression() != null || uim.getActionListeners().length > 0;
+        }
 
         //GET
-        if (menuitem.getUrl() != null || menuitem.getOutcome() != null) {
+        if (isLink) {
             String targetURL = getTargetURL(context, (UIOutcomeTarget) menuitem);
             writer.writeAttribute("href", targetURL, null);
 
@@ -73,9 +81,11 @@ public class MenuItemAwareRenderer extends OutcomeTargetRenderer {
         }
         //POST
         else {
-            writer.writeAttribute("href", "#", null);
-            String menuClientId = source.getClientId(context);
+            writer.writeAttribute("href", "javascript:void(0)", null);
+        }
 
+        if (isCommand) {
+            String menuClientId = source.getClientId(context);
             UIForm form = ComponentTraversalUtils.closestForm(context, source);
             if (form == null) {
                 LOGGER.log(Level.FINE, "Menu '" + menuClientId
@@ -102,6 +112,10 @@ public class MenuItemAwareRenderer extends OutcomeTargetRenderer {
                         : buildNonAjaxRequest(context, ((UIComponent) menuitem), form, ((UIComponent) menuitem).getClientId(context), true);
             }
 
+            if (isLink) {
+                // allow CTRL+CLICK link to open new tab
+                command = "if(PF.metaKey(event)){return true};" + command;
+            }
             onclick = (onclick == null) ? command : onclick + ";" + command;
         }