+++ primefaces/src/main/resources/META-INF/resources/primefaces/core/core.utils.js
@@ -1,4 +1,13 @@
 if (!PrimeFaces.utils) {
+    
+   /**
+    * Shortcut for is this CMD on MacOs or CTRL key on other OSes. 
+    * @param {JQuery.TriggeredEvent} e The key event that occurred.
+    * @return {boolean} `true` if the key is a meta key, or `false` otherwise.
+    */
+    PF.metaKey = function(e) {
+        return PrimeFaces.utils.isMetaKey(e);
+    };
 
     /**
      * The object with various utilities needed by PrimeFaces.
@@ -560,6 +569,15 @@ if (!PrimeFaces.utils) {
             if(e.key === 'Enter') {
                 e.preventDefault();
             }
+        },
+        
+        /**
+         * Is this CMD on MacOs or CTRL key on other OSes. 
+         * @param {JQuery.TriggeredEvent} e The key event that occurred.
+         * @return {boolean} `true` if the key is a meta key, or `false` otherwise.
+         */
+        isMetaKey: function(e) {
+            return PrimeFaces.env.browser.mac ? e.metaKey : e.ctrlKey;
         },
 
         /**