+++ docs/13_0_0/gettingstarted/whatsnew.md
@@ -10,10 +10,18 @@ This page contains a list of big features. Please check the GitHub issues for al
     * Added attribute `primefaces.HTML5_COMPLIANCE` if you know your site is HTML5 doctype so PF won't render certain non-HTML5 compliant values like `text/javascript` on scripts.
 * BlockUI
     * Added attribute `delay` to delay displaying similar to AjaxStatus.
+* Button
+* CommandButton
+* LinkButton
+* MenuButton
+* SplitButton
+* SelectBooleanButton
+    * Added attribute `aria-label` for icon only buttons (will use `title` first if available).
 * Captcha
     * Added attribute `sourceUrl` to override the Google JS location for countries that do not have access to Google.
 * Charts
     * Added attribute `subtitle` exactly the same configuration options with title are available for subtitle.
+    * Added attribute `ariaLabel` to all Charts to allow accessibility but defaults to the Charts title.
 * Column/Columns
     * Added attribute `exportTag` to allow finer grain control over XML export.
 * DataGrid
@@ -30,6 +38,9 @@ This page contains a list of big features. Please check the GitHub issues for al
     * Added attribute `cache` similar to Dialog when `dynamic=true`.
 * OrderList
     * Added attribute `flex` to support PrimeFlex CSS instead of legacy Grid CSS.
+* PickList
+    * Added attribute `filterNormalize` to allow normalized filtering (without accents).
+    * Added attributes `sourceFilterPlaceholder` and `targetFilterPlaceholder` for accessibility.
 * TabView
     * Added `focusOnLastActiveTab` if you want to focus on the tab that the user last activated.
     * Added `footer` facet to add a footer to the whole TabView not per tab.
@@ -44,10 +55,18 @@ This page contains a list of big features. Please check the GitHub issues for al
     * Added support for facet `footer` in overlay panel.
 * SelectManyCheckbox
     * Added attribute `flex` to support PrimeFlex CSS instead of legacy Grid CSS.
+* SelectManyMenu
+    * Added attribute `filterNormalize` to allow normalized filtering (without accents).
+* SelectOneListbox
+    * Added attribute `filterNormalize` to allow normalized filtering (without accents).
+* SelectOneMenu
+    * Added attribute `filterNormalize` to allow normalized filtering (without accents).
 * SelectOneRadio
     * Added attribute `flex` to support PrimeFlex CSS instead of legacy Grid CSS.
 * SpeedDial
     * Added attribute `ariaLabel` to allow screen reader support on button and `title` for tooltip.
+* SplitButton
+    * Added attribute `filterNormalize` to allow normalized filtering (without accents).
 * Menu
     * Added attribute `appendTo` to support custom append of the menu instead of default `@(body)`.
 * MenuItem